requirements.txt
requests
ffmpeg-python
selenium
beautifulsoup4
webdriver-manager